feat: 챌린지 patch API 연동

+++ app/src/main/java/com/greenfriends/zeroway/model/Challenge.kt
@@ -10,7 +10,7 @@ data class ChallengeResponse(
 )
 
 data class ChallengeListResponse(
-    @SerializedName("challengeId") val challengeId: Int,
+    @SerializedName("challengeId") val challengeId: Long,
     @SerializedName("content") val content: String,
     @SerializedName("complete") val complete: Boolean
 )
